requirements.txt
tensorflow
pandas 
dvc
numpy
matplotlib
seaborn
pyYAML
tqdm
joblib
types-PyYAML
scipy
Flask
Flask-Cors
-e .
scikit-learn